0016: 모듈화 준비를 위한 package 정리
