+++ src/main/kotlin/com/smparkworld/discord/common/framework/CommandType.kt
@@ -1,0 +1,12 @@
+package com.smparkworld.discord.common.framework
+
+import com.smparkworld.discord.common.base.StringCode
+import com.smparkworld.discord.common.base.StringsParser.getString
+
+enum class CommandType(
+    val command: String,
+    val description: String
+) {
+    BEE(getString(StringCode.BEE_CMD), getString(StringCode.BEE_CMD_DESC)),
+    VALORANT(getString(StringCode.VAL_CMD), getString(StringCode.VAL_CMD_DESC))
+}
