+++ src/main/kotlin/com/smparkworld/discord/feature/bee/commands/BeeMusicPlayLeaveCommandHandler.kt
@@ -1,0 +1,38 @@
+package com.smparkworld.discord.feature.bee.commands
+
+import com.smparkworld.discord.common.base.StringCode
+import com.smparkworld.discord.common.base.StringsParser.getString
+import com.smparkworld.discord.common.extensions.checkVoiceChannelValidation
+import com.smparkworld.discord.common.extensions.randomValueOf
+import com.smparkworld.discord.common.extensions.requireGuild
+import com.smparkworld.discord.common.extensions.sendNoticeEmbedsMessage
+import com.smparkworld.discord.common.framework.CommandHandler
+import com.smparkworld.discord.domain.GetVoiceChannelByEventAuthorUseCase
+import com.smparkworld.discord.feature.bee.commands.player.MusicManagerMediator
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent
+
+class BeeMusicPlayLeaveCommandHandler(
+    private val musicManagerMediator: MusicManagerMediator,
+    private val getVoiceChannelByEventAuthor: GetVoiceChannelByEventAuthorUseCase,
+) : CommandHandler() {
+
+    override fun handle(command: String, event: SlashCommandInteractionEvent) {
+        checkVoiceChannelValidation(event, result = getVoiceChannelByEventAuthor(event)) { _ ->
+
+            musicManagerMediator.obtainGuildTracker(event.requireGuild().idLong)
+                .scheduler
+                .clearQueue()
+
+            event.requireGuild()
+                .audioManager
+                .closeAudioConnection()
+
+            val message = randomValueOf(
+                getString(StringCode.BEE_CMD_MUSIC_PLAY_LEAVE_MSG_1),
+                getString(StringCode.BEE_CMD_MUSIC_PLAY_LEAVE_MSG_2),
+                getString(StringCode.BEE_CMD_MUSIC_PLAY_LEAVE_MSG_3)
+            )
+            event.sendNoticeEmbedsMessage(message)
+        }
+    }
+}
