+++ src/main/kotlin/com/smparkworld/discord/feature/bee/commands/player/MusicManagerMediator.kt
@@ -1,0 +1,27 @@
+package com.smparkworld.discord.feature.bee.commands.player
+
+import com.sedmelluq.discord.lavaplayer.player.AudioLoadResultHandler
+import com.sedmelluq.discord.lavaplayer.player.DefaultAudioPlayerManager
+import com.sedmelluq.discord.lavaplayer.source.AudioSourceManagers
+import dev.lavalink.youtube.YoutubeAudioSourceManager
+
+private typealias DeprecatedYoutubeSourceManager = com.sedmelluq.discord.lavaplayer.source.youtube.YoutubeAudioSourceManager
+
+class MusicManagerMediator(
+    private val playerManager: DefaultAudioPlayerManager = DefaultAudioPlayerManager(),
+    private val musicManagers: MutableMap<Long, GuildMusicManager> = mutableMapOf()
+) {
+
+    init {
+        playerManager.registerSourceManager(YoutubeAudioSourceManager())
+        AudioSourceManagers.registerRemoteSources(playerManager, DeprecatedYoutubeSourceManager::class.java)
+    }
+
+    fun obtainGuildTracker(guildId: Long): GuildMusicManager {
+        return musicManagers.computeIfAbsent(guildId) { GuildMusicManager(playerManager) }
+    }
+
+    fun loadItem(query: String, handler: AudioLoadResultHandler) {
+        playerManager.loadItem(query, handler)
+    }
+}
