+++ src/main/kotlin/com/smparkworld/discord/feature/valorant/ValorantMapType.kt
@@ -1,0 +1,51 @@
+package com.smparkworld.discord.feature.valorant
+
+enum class ValorantMapType(
+    val typeName: String,
+    val thumbnailUrl: String
+) {
+    SPLIT(
+        typeName = "스플릿",
+        thumbnailUrl = "https://smparkworld.com/media/image/blog/b29056ce-f395-4d9e-b8b9-93efc44f4381.png"
+    ),
+    BIND(
+        typeName = "바인드",
+        thumbnailUrl = "https://smparkworld.com/media/image/blog/2de39087-19b7-4045-8edf-588a62cd8f61.png"
+    ),
+    HAVEN(
+        typeName = "헤이븐",
+        thumbnailUrl = "https://smparkworld.com/media/image/blog/91c207e1-0a34-4cbf-b48f-9b01589444b8.png"
+    ),
+    ASCENT(
+        typeName = "어센트",
+        thumbnailUrl = "https://smparkworld.com/media/image/blog/2d967398-a892-46b3-bf8d-9f11a4f85c8b.png"
+    ),
+    ICEBOX(
+        typeName = "아이스박스",
+        thumbnailUrl = "https://smparkworld.com/media/image/blog/73ff81b9-d220-4c64-9969-93bbe923b1d2.png"
+    ),
+    BREEZE(
+        typeName = "브리즈",
+        thumbnailUrl = "https://smparkworld.com/media/image/blog/fb56fb7b-d85c-41e9-b2bb-bb2d94be7414.png"
+    ),
+    FRACTURE(
+        typeName = "프랙처",
+        thumbnailUrl = "https://smparkworld.com/media/image/blog/a29a10d6-b0a1-459d-a586-d3ad09ada126.png"
+    ),
+    PEARL(
+        typeName = "펄",
+        thumbnailUrl = "https://smparkworld.com/media/image/blog/60b546c2-724a-41d3-83e6-c661c4607b3e.png"
+    ),
+    LOTUS(
+        typeName = "로터스",
+        thumbnailUrl = "https://smparkworld.com/media/image/blog/11027bb3-be47-429a-a584-297cccd519a4.png"
+    ),
+    SUNSET(
+        typeName = "선셋",
+        thumbnailUrl = "https://smparkworld.com/media/image/blog/2f71515d-8013-4b0e-9567-4097b104a82e.png"
+    ),
+    ABYSS(
+        typeName = "어비스",
+        thumbnailUrl = "https://smparkworld.com/media/image/blog/85cee28a-8eeb-4253-b85e-366c81017f91.png"
+    ),
+}
